Made the training process more parametised for ease of work and other QoL stuff
- The network can now train multiple times and puts the results inside a folder
- You can encode/decode input and output data to fit your needs
- You can choose the activation function at runtime
- More losses and accuracy metrics logged
- More information about the network is also present in the logs (topology, learning rate, activation function)
- The prediction on each input (training and validation) is also present in the logs

+++ include/parsing/asm_parser.hpp
@@ -28,16 +28,16 @@ class AsmParser : public IParser {
 	/**
 	 * @brief Parse the input of the neural network into a vector of 0's and 1's according its binary representation
 	 * @param input The string to parse
-	 * @return A vector of doubles representing the input of the neural network
+	 * @return A vector of floats representing the input of the neural network
 	 */
-	std::vector<double> parse_in(std::string& input) override;
+	std::vector<float> parse_in(std::string& input) override;
 
 	/**
 	 * @brief Parse the time taken for the code to run using a header in the first line Cycles: (number)
 	 * @param input The string to parse
-	 * @return A vector of doubles representing the output of the neural network
+	 * @return A vector of floats representing the output of the neural network
 	 */
-	std::vector<double> parse_out(std::string& input) override;
+	std::vector<float> parse_out(std::string& input) override;
 };
 
 #endif // ASM_PARSER_HPP
