+++ include/parsing/binary_parser.hpp
@@ -28,16 +28,16 @@ class BinaryParser : public IParser {
 	/**
 	 * @brief Parses a binary representation of assembly code into a vector of 0's and 1's according to its binary representation,
 	 * @param input The binary code to parse
-	 * @return A vector of doubles representing the input of the neural network
+	 * @return A vector of floats representing the input of the neural network
 	 */
-	std::vector<double> parse_in(std::string& input) override;
+	std::vector<float> parse_in(std::string& input) override;
 
 	/**
 	 * @brief Parse the time taken for the code to run using a header in the first line Cycles: (number)
 	 * @param input The string to parse
-	 * @return A vector of doubles representing the output of the neural network
+	 * @return A vector of floats representing the output of the neural network
 	 */
-	std::vector<double> parse_out(std::string& input) override;
+	std::vector<float> parse_out(std::string& input) override;
 };
 
 #endif // BINARY_PARSER_HPP
