+++ src/parsing/xor_parser.cpp
@@ -3,15 +3,12 @@
  * @brief Implementation of the XORParser class
  */
 
-#include <fstream>
-#include <iostream>
-#include <tuple>
 #include <vector>
 
 #include "parsing/xor_parser.hpp"
 
-std::vector<double> XORParser::parse_in(std::string& input) {
-	std::vector<double> res = {0.0, 0.0};
+std::vector<float> XORParser::parse_in(std::string& input) {
+	std::vector<float> res = {0.0, 0.0};
 	if (input.length() < 3) {
 		return res;
 	}
@@ -24,8 +21,8 @@ std::vector<double> XORParser::parse_in(std::string& input) {
 	return res;
 }
 
-std::vector<double> XORParser::parse_out(std::string& input) {
-	std::vector<double> res = {0.0};
+std::vector<float> XORParser::parse_out(std::string& input) {
+	std::vector<float> res = {0.0};
 	if (input.length() < 5) {
 		return res;
 	}
